% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.r
\name{hello}
\alias{hello}
\title{A first function in an example package}
\usage{
hello(name)
}
\arguments{
\item{name}{the name of a person to greet.}
}
\value{
A string greeting a person.
}
\description{
This is the first function of the p1 package
written.
}
\examples{

print(hello("bis557"))

}
